feature/Dobee work with windw object then helper.
