+++ lib/Domino.js
@@ -1,9 +1,21 @@
-import { setCallback, generateProps } from "../util/helper";
+import Dobee from "./Dobee";
+
+function getProps(propsText) {
+  if(!propsText) return {};
+
+  const props = JSON.parse(propsText);
+  Object.keys(props).forEach((key) => {
+    if(key.startsWith("on")) {
+      props[key] = Dobee.getCallback(key);
+    }
+  });
+  return props;
+}
 
 export default class Component extends HTMLElement {
   constructor() {
     super();
-    this._props = generateProps(this.getAttribute("props"));
+    this._props = getProps(this.getAttribute("props"));
     this.eventList = {};
   }
 
@@ -13,7 +25,7 @@ export default class Component extends HTMLElement {
   }
 
   useCallback(key, callback) {
-    return setCallback(key, callback);
+    return Dobee.setCallback(key, callback);
   }
 
   useEvent(query, type, callback) {
@@ -50,12 +62,10 @@ export default class Component extends HTMLElement {
     this.setEvent();
   }
 
+  useProps(props) {
+    return JSON.stringify(props);
+  }
+
   adoptCallback() {} 
   attributeChangedCallback(attrName, oldVal, newVal) {}
 }
-
-function useProps(props) {
-  return JSON.stringify(props);
-}
-
-export {useProps};
